# server/api/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Environment Manager API
host: environmentmanager.corp.local
basePath: /api/v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /environments/{name}:
    x-swagger-router-controller: environmentsController
    get:
      operationId: getEnvironmentByName
      summary: Get an environment.
      tags:
        - Environment
      parameters:
        - name: name
          in: path
          type: string
          required: true
      responses:
        200:
          description: The environment
          schema:
            $ref: '#/definitions/OpsEnvironment'
...
  /config/deployment-maps:
    x-swagger-router-controller: deploymentMapController
    get:
      operationId: getDeploymentMapsConfig
      summary: Get all deployment map configurations.
      tags:
        - Deployment Map
      responses:
        200:
          description: The list of deployment maps
          schema:
            type: array
            items:
              $ref: '#/definitions/DeploymentMap'
    post:
      operationId: postDeploymentMapsConfig
      summary: Create a deployment map configuration
      tags:
        - Deployment Map
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/DeploymentMap'
      responses:
        201:
          description: OK
  /config/deployment-maps/{name}:
    x-swagger-router-controller: deploymentMapController
    put:
      operationId: putDeploymentMapConfigByName
      summary: Update a deployment map configuration.
      tags:
        - Deployment Map
      parameters:
        - name: name
          in: path
          type: string
          required: true
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/DeploymentMapValue'
        - $ref: '#/parameters/ExpectedVersionHeader'
      responses:
        200:
          description: OK
    get:
      operationId: getDeploymentMapConfigByName
      summary: Get a specific deployment map configuration.
      tags:
        - Deployment Map
      parameters:
        - name: name
          in: path
          type: string
          required: true
      responses:
        200:
          description: The deployment map config
          schema:
            $ref: '#/definitions/DeploymentMap'
    delete:
      operationId: deleteDeploymentMapConfigByName
      summary: Delete a deployment map configuration.
      tags:
        - Deployment Map
      parameters:
        - name: name
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /config/environments:
    x-swagger-router-controller: environmentsConfigController
    get:
      operationId: getEnvironmentsConfig
      summary: Get all environment configurations.
      tags:
        - Environment
      parameters:
        - name: environmentType
          in: query
          type: string
          required: false
        - name: cluster
          in: query
          required: false
          type: string
      responses:
        200:
          description: The list of environment configs
          schema:
            type: array
            items:
              $ref: '#/definitions/EnvironmentConfig'
    post:
      operationId: postEnvironmentsConfig
      summary: Create a new environment configuration.
      tags:
        - Environment
      x-authorizer: environments
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/EnvironmentConfig'
      responses:
        201:
          description: OK
  /config/environments/{name}:
    x-swagger-router-controller: environmentsConfigController
    get:
      operationId: getEnvironmentConfigByName
      summary: Get a specific environment configuration.
      tags:
        - Environment
      parameters:
        - name: name
          in: path
          type: string
          required: true
      responses:
        200:
          description: The environment config
          schema:
            type: object
    put:
      operationId: putEnvironmentConfigByName
      summary: Update an environment configuration.
      tags:
        - Environment
      x-authorizer: environments
      parameters:
        - name: name
          in: path
          type: string
          required: true
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/EnvironmentConfigValue'
        - $ref: '#/parameters/ExpectedVersionHeader'
      responses:
        200:
          description: OK
    delete:
      operationId: deleteEnvironmentConfigByName
      summary: Delete an environment configuration.
      tags:
        - Environment
      x-authorizer: environments
      parameters:
        - name: name
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /config/environment-types:
    x-swagger-router-controller: environmentTypeController
    get:
      operationId: getEnvironmentTypesConfig
      summary: Get all environment type configurations.
      tags:
        - Environment Type
      responses:
        200:
          description: The list of environment type configs
          schema:
            type: array
            items:
              $ref: '#/definitions/EnvironmentTypeConfig'
    post:
      operationId: postEnvironmentTypesConfig
      summary: Create an Environment Type configuration.
      tags:
        - Environment Type
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/EnvironmentTypeConfig'
      responses:
        201:
          description: OK
  /config/environment-types/{name}:
    x-swagger-router-controller: environmentTypeController
    get:
      operationId: getEnvironmentTypeConfigByName
      summary: Get an specific environment type configuration.
      tags:
        - Environment Type
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        200:
          description: The environment type config
          schema:
            type: object
    put:
      operationId: putEnvironmentTypeConfigByName
      summary: Update an environment type configuration.
      tags:
        - Environment Type
      parameters:
        - name: name
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/EnvironmentTypeConfigValue'
        - $ref: '#/parameters/ExpectedVersionHeader'
      responses:
        200:
          description: OK
    delete:
      operationId: deleteEnvironmentTypeConfigByName
      summary: Delete an environment type.
      tags:
        - Environment Type
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        200:
          description: OK
...
definitions:
...
  DeploymentMap:
    type: object
    required:
      - DeploymentMapName
      - Value
    properties:
      DeploymentMapName:
        type: string
      Value:
        $ref: '#/definitions/DeploymentMapValue'
  DeploymentMapValue:
    type: object
    properties:
      DeploymentTarget:
        type: array
        items:
          $ref: '#/definitions/DeploymentTarget'
      SchemaVersion:
        type: integer
  DeploymentTarget:
    type: object
    properties:
      AMI:
        type: string
      AutoScalingSettings:
        type: object
        properties:
          DesiredCapacity:
            type: integer
          MaxCapacity:
            type: integer
          MinCapacity:
            type: integer
      ContactEmailTag:
        type: string
      FleetPerSlice:
        type: boolean
      InstanceType:
        type: string
      OwningCluster:
        type: string
      RoleTag:
        type: string
      SecurityZone:
        type: string
      ServerRoleName:
        type: string
      Services:
        type: array
        items:
          type: object
          properties:
            ServiceName:
              type: string
      SubnetTypeName:
        type: string
      Volumes:
        type: array
        items:
          type: object
          properties:
            Name:
              type: string
            Size:
              type: integer
            Type:
              type: string
  OpsEnvironment:
    type: object
    required:
      - EnvironmentName
      - Value
    properties:
      EnvironmentName:
        type: string
      Value:
        $ref: '#/definitions/OpsEnvironmentValue'
  OpsEnvironmentValue:
    type: object
    properties:
      AccountName:
        type: string
      ScheduleStatus:
        type: string
...
  EnvironmentConfig:
    type: object
    required:
      - EnvironmentName
      - Value
    properties:
      EnvironmentName:
        type: string
      AWSAccountNumber:
        type: string
      Version:
        type: integer
      Value:
        $ref: '#/definitions/EnvironmentConfigValue'
  EnvironmentConfigValue:
    type: object
    properties:
      OwningCluster:
        type: string
      AlertSettings:
        type: string
        enum:
          - EnvironmentOwner
          - ServiceOwner
          - Custom
      NotificationSettingsId:
        type: string
      Description:
        type: string
      EnvironmentType:
        type: string
      SchemaVersion:
        type: integer
      DeploymentMap:
        type: string
  EnvironmentTypeConfig:
    type: object
    required:
      - EnvironmentType
      - Value
    properties:
      EnvironmentType:
        type: string
      Value:
        $ref: '#/definitions/EnvironmentTypeConfigValue'
  EnvironmentTypeConfigValue:
    type: object
    properties:
      AWSAccountName:
        type: string
      AWSAccountNumber:
        type: string
      Consul:
        type: object
        properties:
          DataCenter:
            type: string
          Port:
            type: integer
          SecurityTokenPath:
            type: object
            properties:
              Bucket:
                type: string
              Key:
                type: string
          Servers:
            type: array
            items:
              type: string
      DeploymentBucket:
        type: string
      LoadBalancers:
        type: array
        items:
          type: string
      SchemaVersion:
        type: integer
      Subnets:
        type: object
      VpcId:
        type: string
...
parameters:
  ExpectedVersionHeader:
    name: expected-version
    in: header
    required: true
    type: integer
...